+++ crates/krb5-types/src/lib.rs
@@ -30,6 +30,7 @@ pub mod pkinit;
 pub mod s4u;
 pub mod spake;
 pub mod timestamp;
+pub mod transited;
 
 pub use constants::{ap_bit, err, flag_bit, ku, pa};
 pub use extra::{